Add restrictions for users

+++ cafe/views.py
@@ -65,19 +65,20 @@ def edit_specials(request, special_id):
     Enables superuser to edit items in
     today's specials database.
     """
-    item = get_object_or_404(Specials, id=special_id)
-    if request.method == "POST":
-        form = SpecialsForm(request.POST, instance=item)
-        if form.is_valid():
-            specials = form.save()
-            specials.save()
-            messages.success(request, 'An item has been updated.')
-        return redirect('today_specials')
-    form = SpecialsForm(instance=item)
-    context = {
-        'form': form
-    }
-    return render(request, 'edit_specials.html', context)
+    if request.user.is_superuser:
+        item = get_object_or_404(Specials, id=special_id)
+        if request.method == "POST":
+            form = SpecialsForm(request.POST, instance=item)
+            if form.is_valid():
+                specials = form.save()
+                specials.save()
+                messages.success(request, 'An item has been updated.')
+            return redirect('today_specials')
+        form = SpecialsForm(instance=item)
+        context = {
+            'form': form
+        }
+        return render(request, 'edit_specials.html', context)
 
 
 @login_required
@@ -86,18 +87,19 @@ def delete_specials(request, special_id):
     Enables superuser to delete items in
     today's specials database.
     """
-    item = get_object_or_404(Specials, id=special_id)
-    if request.method == "POST":
-        form = SpecialsForm(request.POST, instance=item)
-        if item.delete():
-            messages.success(request, 'The item has been deleted.')
-            return redirect('today_specials')
+    if request.user.is_superuser:
+        item = get_object_or_404(Specials, id=special_id)
+        if request.method == "POST":
+            form = SpecialsForm(request.POST, instance=item)
+            if item.delete():
+                messages.success(request, 'The item has been deleted.')
+                return redirect('today_specials')
 
-    form = SpecialsForm(instance=item)
-    context = {
-        'form': form
-    }
-    return render(request, 'delete_specials.html', context)
+        form = SpecialsForm(instance=item)
+        context = {
+            'form': form
+        }
+        return render(request, 'delete_specials.html', context)
 
 
 @login_required
